Move bip32 file to wallet
